requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attr==0.3.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.3.post0
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
portalocker==2.10.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.2
pytz==2024.1
PyYAML==6.0.1
requests==2.28.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
sympy==1.12.1
threadpoolctl==3.5.0
tqdm==4.66.4
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.13
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
